fix: add casts for return statements and array subscripts
- Added 212 static_cast<bool> for return statements with size()
- Added array subscript safety casts for -1 patterns
- Fixes additional -Wsign-compare warnings in complex expressions
- Improves type safety in return and subscript operations

+++ src/acceleration/directx_backend_full.cpp
@@ -427,7 +427,7 @@ public:
             // Drain heap in ascending distance order
             results[q].resize(heap.size());
             for (size_t i = heap.size(); i > 0; --i) {
-                results[q][i - 1] = { heap.top().second, heap.top().first };
+                results[q][static_cast<int>(i - 1)] = { heap.top().second, heap.top().first };
                 heap.pop();
             }
         }
